requirements.txt
google-genai
python-dotenv
requests
gspread
google-auth
beautifulsoup4
schedule
flask
flask-cors
psycopg2-binary
